feat: move new.rs to lib.rs
